修改文件上传大小限制200MB

+++ src/main/resources/db/migration/V20220307210904__create_file_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE file
     `name`         VARCHAR(64)   NOT NULL COMMENT '文件名',
     `file_key`     VARCHAR(64)   NOT NULL COMMENT '文件hash值，即ObjectKey',
     `ext`          VARCHAR(12)   NOT NULL COMMENT '文件后缀名',
-    `size`         INT DEFAULT 0 NOT NULL COMMENT '文件大小；单位byte',
+    `size`         BIGINT DEFAULT 0 NOT NULL COMMENT '文件大小；单位byte',
     `type`         VARCHAR(32)   NULL COMMENT '文件类型，MP3-音频，IMAGE-图片，LYRIC-歌词',
     `storage_url`  VARCHAR(128)  NOT NULL COMMENT '文件在OSS中存储位置，用于oss增删改查',
     `visit_url`    VARCHAR(128)  NOT NULL COMMENT '文件公网访问url',
